zsh: Fix PATH on macOS
The macOS base system evals path_helper in /etc/zprofile, which is
loaded after zshenv. Since zshenv should be used for setting most
environment variables, this is a mistake, and causes configuration in
zshenv that tests for command existence to sometimes miss the presence
of commands.

Workaround the base system error by manually evaling path_helper.