shell: Alias dnf to always pass --cacheonly
Many dnf commands can run without elevated privileges, but the default
behavior is to make a separate copy of the cache for that user. This is
slow and means multiple caches need to be maintained. Passing
`--cache-only` means that it will just use the existing system cache
without updating.

Since this behavior is only desired when not root, don’t add this alias
as root.