bin/git-branch-cleanup: Cleanup branches orphaned from their remote
That is: branches which have an upstream tracking branch specified but
that upstream tracking branch has been deleted from the remote.

We don’t `git fetch` in this scenario since it seems unnecessary. If the
remote ref is still present locally but really gone, then cleanup is
just deferred till later (seems fine). If the local ref is gone but
exists on the remote then it *was* gone and has reappeared. This seems
unlikely in most cases and even then the new remote ref may be unrelated
to the local.