git: Fix recursively fetching submodules
Elsewhere we set fetch.recurseSubmodules in an attempt to override
submodule.recurse, however that did not seem to have any effect. This
appears to be because Git parses config in-order, and `git-fetch` reads
both of these and the last one wins.

As such, move the higher-level one earlier in the config to ensure it
gets overwritten instead.